Agent/master_agent/requirements.txt
google-cloud-aiplatform[adk,agent_engines]
Flask>=2.3.0
requests
python-dotenv
